Started moving the authentication to the core layer #37 Also changed the way we load the modules for the IoC container, we just need to create a class inside the DependencyResolver.Module assembly that is derived from NinjectModule.
