+++ NZBDash.Core/NZBDash.Core.csproj
@@ -107,6 +107,10 @@
     <Reference Include="EntityFramework.SqlServer">
       <HintPath>..\packages\EntityFramework.6.1.3\lib\net45\EntityFramework.SqlServer.dll</HintPath>
     </Reference>
+    <Reference Include="Microsoft.AspNet.Identity.Core, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\packages\Microsoft.AspNet.Identity.Core.2.1.0\lib\net45\Microsoft.AspNet.Identity.Core.dll</HintPath>
+    </Reference>
     <Reference Include="Microsoft.VisualBasic" />
     <Reference Include="Microsoft.VisualBasic.Compatibility" />
     <Reference Include="Microsoft.VisualBasic.Compatibility.Data" />
@@ -140,6 +144,7 @@
     <Compile Include="Model\Settings\SabNzbdSettingsDto.cs" />
     <Compile Include="Model\Settings\SonarrSettingsViewModelDto.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\AuthenticationService.cs" />
     <Compile Include="Services\MonoHardwareService.cs" />
     <Compile Include="Services\WindowsHardwareService.cs" />
     <Compile Include="SettingsService\NzbDashSettingsService.cs" />
@@ -178,10 +183,6 @@
     <ProjectReference Include="..\NZBDash.DataAccessLayer.Interfaces\NZBDash.DataAccessLayer.Interfaces.csproj">
       <Project>{A1BAFD2C-9DF1-4A1A-A6CA-0DF2C2E60DFC}</Project>
       <Name>NZBDash.DataAccessLayer.Interfaces</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\NZBDash.DependencyResolver.ServiceLocator\NZBDash.DependencyResolver.ServiceLocator.csproj">
-      <Project>{BC304B86-E724-4953-AEA9-9541D175E51D}</Project>
-      <Name>NZBDash.DependencyResolver.ServiceLocator</Name>
     </ProjectReference>
     <ProjectReference Include="..\NZBDash.ThirdParty.Api.Models\NZBDash.ThirdParty.Api.Models.csproj">
       <Project>{E51F0F88-A6DA-4A7E-B647-C5621B60F79D}</Project>
