+++ NZBDash.sln
@@ -56,8 +56,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NZBDash.Common.Helpers", "N
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NZBDash.DataAccessLayer.Models", "NZBDash.DataAccessLayer.Models\NZBDash.DataAccessLayer.Models.csproj", "{67E1FFF8-EDEC-4D4E-8B49-E3B107D326B6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NZBDash.DependencyResolver.ServiceLocator", "NZBDash.DependencyResolver.ServiceLocator\NZBDash.DependencyResolver.ServiceLocator.csproj", "{BC304B86-E724-4953-AEA9-9541D175E51D}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -736,30 +734,6 @@ Global
 		{67E1FFF8-EDEC-4D4E-8B49-E3B107D326B6}.Windows|Mixed Platforms.Build.0 = Release|Any CPU
 		{67E1FFF8-EDEC-4D4E-8B49-E3B107D326B6}.Windows|x64.ActiveCfg = Release|Any CPU
 		{67E1FFF8-EDEC-4D4E-8B49-E3B107D326B6}.Windows|x86.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Linux|Any CPU.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Linux|Any CPU.Build.0 = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Linux|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Linux|Mixed Platforms.Build.0 = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Linux|x64.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Linux|x86.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Release|x64.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Release|x86.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Windows|Any CPU.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Windows|Any CPU.Build.0 = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Windows|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Windows|Mixed Platforms.Build.0 = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Windows|x64.ActiveCfg = Release|Any CPU
-		{BC304B86-E724-4953-AEA9-9541D175E51D}.Windows|x86.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
